{"schema":"decodex/commit/1","summary":"Split plugin, MCP, and app server test modules","authority":"manual"}
